faction blending

+++ RVCRestructured/Source/VineLib/RVR/Harmony/Apparel/ScoreGain.cs
@@ -9,7 +9,7 @@ public static class ApparelScoreGainPatch
     {
         if (!pawn.CanUse(ap.def))
         {
-            __result = int.MinValue;
+            __result = float.MinValue;
             return;
         }
     }
